LayoutTests/platform/chromium-win/fast/events/onunload-not-on-body-expected.txt

frame "<!--framePath //<!--frame2-->-->" - has 1 onunload handler(s)
ALERT: unload
you should only see one unload alert appear.
